[deckhouse] move the startup sync into internal/controller/syncer
- One file per synced resource: `module_package_version.go`, `module_package.go`; a future resource gets its own file.
- The syncer lives beside the controller bootstrap that calls it.
- The log source is named "syncer".
