+++ deckhouse-controller/internal/controller/syncer/module_package_version.go
@@ -12,37 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package versionsync creates the ModulePackageVersion and ModulePackage
-// objects for the module packages the old module stack already carries in the
-// cluster, so the package system sees the versions the cluster runs and the
-// packages its sources offer.
-//
-// # Data sources
-//
-//	embedded modules dir (the running image)
-//	  └─ embedded-<module>-<deckhouse version>, complete: the metadata is
-//	     filled from the module files on disk, no repository ever serves it
-//
-//	deployed or pending ModuleRelease
-//	  └─ <repository>-<module>-<version>, where the "deckhouse" source maps
-//	     to the "deckhouse-modules" repository; a draft stub - the
-//	     module-package-version controller fills it once a PackageRepository
-//	     exists
-//
-//	v1alpha1 Module with a non-empty availableSources
-//	  └─ ModulePackage <module>: availableRepositories carries the sources
-//	     mapped to repository names; the repository scan later adopts the
-//	     package and appends the repositories it really serves
-//
-// A version stays a draft until its metadata lands, so no observer takes a
-// half-created version for a complete one; a fill interrupted mid-way heals on
-// the next start. The legacy label keeps the registry path of the module
-// source world ("<module>/release"). No owner is set: the repository the spec
-// names may not exist yet, and an owner reference to a missing object would
-// get the version garbage-collected; the repository scan adopts the stubs it
-// recognizes. An existing complete version is never touched, so a restart
-// changes nothing.
-package versionsync
+package syncer
 
 import (
 	"context"
@@ -64,55 +34,8 @@ import (
 	"github.com/deckhouse/deckhouse/deckhouse-controller/internal/app"
 	"github.com/deckhouse/deckhouse/deckhouse-controller/internal/packages/loader"
 	"github.com/deckhouse/deckhouse/deckhouse-controller/pkg/apis/deckhouse.io/v1alpha1"
-	"github.com/deckhouse/deckhouse/go_lib/dependency"
 	"github.com/deckhouse/deckhouse/pkg/log"
 )
-
-// Syncer creates the missing package versions once at start, while the
-// controllers still wait for the sync phase.
-type Syncer struct {
-	// reader must bypass the manager cache: the ModulePackageVersion kind is
-	// cached only when the module packages feature is on, and this sync runs
-	// everywhere.
-	reader client.Reader
-	writer client.Client
-	dc     dependency.Container
-
-	deckhouseVersion   string
-	embeddedModulesDir string
-
-	logger *log.Logger
-}
-
-// New builds a Syncer for the given Deckhouse version and embedded modules dir.
-func New(reader client.Reader, writer client.Client, dc dependency.Container, deckhouseVersion, embeddedModulesDir string, logger *log.Logger) *Syncer {
-	return &Syncer{
-		reader: reader,
-		writer: writer,
-		dc:     dc,
-
-		deckhouseVersion:   deckhouseVersion,
-		embeddedModulesDir: embeddedModulesDir,
-
-		logger: logger,
-	}
-}
-
-// Sync ensures the package versions of the old module stack. A source naming
-// no valid version (no module source, an unparsable version, an illegal object
-// name, an unreadable module dir) is skipped with a warning; an API failure
-// stops the sync.
-func (s *Syncer) Sync(ctx context.Context) error {
-	if err := s.syncEmbedded(ctx); err != nil {
-		return err
-	}
-
-	if err := s.syncReleases(ctx); err != nil {
-		return err
-	}
-
-	return s.syncPackages(ctx)
-}
 
 // syncEmbedded walks the embedded modules dir and ensures a complete version
 // for every module the running image ships.
@@ -387,102 +310,6 @@ func (s *Syncer) removeDraft(ctx context.Context, mpv *v1alpha1.ModulePackageVer
 	}
 
 	s.logger.Debug("module package version filled from disk", slog.String("name", mpv.Name))
-
-	return nil
-}
-
-// syncPackages ensures a ModulePackage for every module the old module sources
-// offer, so the catalog knows where each package is available before any
-// repository scan ran.
-func (s *Syncer) syncPackages(ctx context.Context) error {
-	modules := new(v1alpha1.ModuleList)
-	if err := s.reader.List(ctx, modules); err != nil {
-		return fmt.Errorf("list modules: %w", err)
-	}
-
-	for idx := range modules.Items {
-		module := &modules.Items[idx]
-
-		// an embedded module comes from no source; the scan never creates a
-		// package for it either
-		if module.IsEmbedded() || len(module.Properties.AvailableSources) == 0 {
-			continue
-		}
-
-		if err := s.ensurePackage(ctx, module.Name, repositoriesForSources(module.Properties.AvailableSources)); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// repositoriesForSources maps module source names onto repository names,
-// dropping the duplicates the mapping may fold together.
-func repositoriesForSources(sources []string) []string {
-	repositories := make([]string, 0, len(sources))
-
-	for _, source := range sources {
-		repository := v1alpha1.PackageRepositoryNameForModuleSource(source)
-		if !slices.Contains(repositories, repository) {
-			repositories = append(repositories, repository)
-		}
-	}
-
-	return repositories
-}
-
-// ensurePackage makes sure the package exists with the given repositories in
-// its status. An existing package is left untouched unless its repository list
-// is empty, which a create interrupted before the status patch leaves behind;
-// such a package is completed in place. No owner is set: the repositories the
-// status names may not exist yet, and the scan adopts the package once one
-// appears.
-func (s *Syncer) ensurePackage(ctx context.Context, name string, repositories []string) error {
-	pkg := new(v1alpha1.ModulePackage)
-
-	err := s.reader.Get(ctx, client.ObjectKey{Name: name}, pkg)
-	if err != nil && !apierrors.IsNotFound(err) {
-		return fmt.Errorf("get module package '%s': %w", name, err)
-	}
-
-	if apierrors.IsNotFound(err) {
-		pkg = &v1alpha1.ModulePackage{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: v1alpha1.ModulePackageGVK.GroupVersion().String(),
-				Kind:       v1alpha1.ModulePackageKind,
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:   name,
-				Labels: map[string]string{"heritage": "deckhouse"},
-			},
-		}
-
-		if err := s.writer.Create(ctx, pkg); err != nil {
-			if !apierrors.IsAlreadyExists(err) {
-				return fmt.Errorf("create module package '%s': %w", name, err)
-			}
-
-			// another writer created it between the read and this call;
-			// converge on whatever exists now
-			if err := s.reader.Get(ctx, client.ObjectKey{Name: name}, pkg); err != nil {
-				return fmt.Errorf("get module package '%s': %w", name, err)
-			}
-		}
-	}
-
-	if len(pkg.Status.AvailableRepositories) > 0 {
-		return nil
-	}
-
-	original := pkg.DeepCopy()
-	pkg.Status.AvailableRepositories = repositories
-
-	if err := s.writer.Status().Patch(ctx, pkg, client.MergeFrom(original)); err != nil {
-		return fmt.Errorf("patch module package status '%s': %w", name, err)
-	}
-
-	s.logger.Debug("module package created", slog.String("name", name))
 
 	return nil
 }
