+++ deckhouse-controller/internal/controller/syncer/module_package_version_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package versionsync
+package syncer
 
 import (
 	"context"
@@ -24,82 +24,9 @@ import (
 	"github.com/stretchr/testify/require"
 	metautils "k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/deckhouse/deckhouse/deckhouse-controller/pkg/apis/deckhouse.io/v1alpha1"
-	"github.com/deckhouse/deckhouse/go_lib/dependency"
-	"github.com/deckhouse/deckhouse/go_lib/project"
-	"github.com/deckhouse/deckhouse/pkg/log"
 )
-
-func newTestSyncer(t *testing.T, version, embeddedDir string, objects ...client.Object) (*Syncer, client.Client) {
-	t.Helper()
-
-	sc, err := project.Scheme()
-	require.NoError(t, err)
-
-	cl := fake.NewClientBuilder().
-		WithScheme(sc).
-		WithStatusSubresource(&v1alpha1.ModulePackageVersion{}, &v1alpha1.ModulePackage{}).
-		WithObjects(objects...).
-		Build()
-
-	return New(cl, cl, dependency.NewMockedContainer(), version, embeddedDir, log.NewNop()), cl
-}
-
-func writeModuleYAML(t *testing.T, dir, content string) {
-	t.Helper()
-	require.NoError(t, os.MkdirAll(dir, 0o755))
-	require.NoError(t, os.WriteFile(filepath.Join(dir, "module.yaml"), []byte(content), 0o644))
-}
-
-func writePackageYAML(t *testing.T, dir, content string) {
-	t.Helper()
-	require.NoError(t, os.MkdirAll(dir, 0o755))
-	require.NoError(t, os.WriteFile(filepath.Join(dir, "package.yaml"), []byte(content), 0o644))
-}
-
-func testModule(name, source string, available ...string) *v1alpha1.Module {
-	return &v1alpha1.Module{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Properties: v1alpha1.ModuleProperties{Source: source, AvailableSources: available},
-	}
-}
-
-func testRelease(module, source, version, phase string) *v1alpha1.ModuleRelease {
-	return &v1alpha1.ModuleRelease{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   module + "-v" + version,
-			Labels: map[string]string{"source": source},
-		},
-		Spec:   v1alpha1.ModuleReleaseSpec{ModuleName: module, Version: version},
-		Status: v1alpha1.ModuleReleaseStatus{Phase: phase},
-	}
-}
-
-func listVersionNames(t *testing.T, cl client.Client) []string {
-	t.Helper()
-
-	list := new(v1alpha1.ModulePackageVersionList)
-	require.NoError(t, cl.List(context.Background(), list))
-
-	names := make([]string, 0, len(list.Items))
-	for _, item := range list.Items {
-		names = append(names, item.Name)
-	}
-
-	return names
-}
-
-func getVersion(t *testing.T, cl client.Client, name string) *v1alpha1.ModulePackageVersion {
-	t.Helper()
-
-	mpv := new(v1alpha1.ModulePackageVersion)
-	require.NoError(t, cl.Get(context.Background(), client.ObjectKey{Name: name}, mpv))
-
-	return mpv
-}
 
 func TestSyncEmbedded(t *testing.T) {
 	ctx := context.Background()
@@ -293,101 +220,4 @@ func TestSyncReleases(t *testing.T) {
 		assert.Equal(t, before.ResourceVersion, after.ResourceVersion)
 		assert.False(t, after.IsDraft(), "a complete version must not become a draft")
 	})
-}
-
-func getPackage(t *testing.T, cl client.Client, name string) *v1alpha1.ModulePackage {
-	t.Helper()
-
-	pkg := new(v1alpha1.ModulePackage)
-	require.NoError(t, cl.Get(context.Background(), client.ObjectKey{Name: name}, pkg))
-
-	return pkg
-}
-
-func TestSyncPackages(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("creates a package for every module the sources offer", func(t *testing.T) {
-		s, cl := newTestSyncer(t, "v1.80.0", t.TempDir(),
-			testModule("parca", "deckhouse", "deckhouse"),
-			testModule("foo", "external", "deckhouse", "external"),
-			testModule("echo", v1alpha1.ModuleSourceEmbedded),
-			testModule("bare", ""),
-		)
-
-		require.NoError(t, s.Sync(ctx))
-
-		list := new(v1alpha1.ModulePackageList)
-		require.NoError(t, cl.List(ctx, list))
-		require.Len(t, list.Items, 2, "embedded and sourceless modules name no package")
-
-		parca := getPackage(t, cl, "parca")
-		assert.Equal(t, []string{"deckhouse-modules"}, parca.Status.AvailableRepositories, "the source maps to the repository name")
-		assert.Equal(t, "deckhouse", parca.Labels["heritage"])
-		assert.Empty(t, parca.OwnerReferences, "no owner until a repository adopts the package")
-
-		foo := getPackage(t, cl, "foo")
-		assert.Equal(t, []string{"deckhouse-modules", "external"}, foo.Status.AvailableRepositories)
-	})
-
-	t.Run("keeps an existing package untouched", func(t *testing.T) {
-		existing := &v1alpha1.ModulePackage{
-			ObjectMeta: metav1.ObjectMeta{Name: "parca"},
-			Status:     v1alpha1.ModulePackageStatus{AvailableRepositories: []string{"other-repo"}},
-		}
-
-		s, cl := newTestSyncer(t, "v1.80.0", t.TempDir(), existing,
-			testModule("parca", "deckhouse", "deckhouse"),
-		)
-		before := getPackage(t, cl, existing.Name)
-
-		require.NoError(t, s.Sync(ctx))
-
-		after := getPackage(t, cl, existing.Name)
-		assert.Equal(t, before.ResourceVersion, after.ResourceVersion)
-		assert.Equal(t, []string{"other-repo"}, after.Status.AvailableRepositories)
-	})
-
-	t.Run("completes a package left without repositories", func(t *testing.T) {
-		interrupted := &v1alpha1.ModulePackage{ObjectMeta: metav1.ObjectMeta{Name: "parca"}}
-
-		s, cl := newTestSyncer(t, "v1.80.0", t.TempDir(), interrupted,
-			testModule("parca", "deckhouse", "deckhouse"),
-		)
-
-		require.NoError(t, s.Sync(ctx))
-
-		assert.Equal(t, []string{"deckhouse-modules"}, getPackage(t, cl, "parca").Status.AvailableRepositories,
-			"a create interrupted before the status patch must be completed")
-	})
-}
-
-func TestSyncIsIdempotent(t *testing.T) {
-	ctx := context.Background()
-
-	dir := t.TempDir()
-	writeModuleYAML(t, filepath.Join(dir, "900-echo"), "name: echo\nstage: General Availability\n")
-
-	s, cl := newTestSyncer(t, "v1.80.0", dir,
-		testRelease("parca", "deckhouse", "1.4.3", v1alpha1.ModuleReleasePhaseDeployed),
-		testRelease("console", "deckhouse", "1.60.1", v1alpha1.ModuleReleasePhasePending),
-		testModule("parca", "deckhouse", "deckhouse"),
-	)
-
-	require.NoError(t, s.Sync(ctx))
-
-	versions := make(map[string]string)
-	for _, name := range listVersionNames(t, cl) {
-		versions[name] = getVersion(t, cl, name).ResourceVersion
-	}
-	require.Len(t, versions, 3)
-	packageRV := getPackage(t, cl, "parca").ResourceVersion
-
-	require.NoError(t, s.Sync(ctx))
-
-	assert.Len(t, listVersionNames(t, cl), 3)
-	for name, rv := range versions {
-		assert.Equal(t, rv, getVersion(t, cl, name).ResourceVersion, name)
-	}
-	assert.Equal(t, packageRV, getPackage(t, cl, "parca").ResourceVersion)
 }
